Script for SA on 3 parameters of interest

+++ modelling/scripts/SA_drug.py
@@ -1,34 +1,15 @@
 # To plot the range of parameter values
 # Drug binding-related parameters
 
-import matplotlib.pyplot as plt
 import myokit
 import numpy as np
 import os
 import pandas as pd
+import pints
 
 import modelling
 
 saved_data_dir = '../../simulation_data/sensitivity_analysis/'
-
-testing_fig_dir = '../../figures/testing/'
-final_fig_dir = '../../figures/binding_kinetics_comparison/' + \
-    'OHaraCiPA_model/sensitivity_analysis/'
-
-check_plot = True
-saved_fig_dir = final_fig_dir
-
-param_lib = modelling.BindingParameters()
-drug_list = param_lib.drug_compounds
-
-drug = 'dofetilide'
-Vhalf = param_lib.binding_parameters[drug]['Vhalf']
-Kmax = param_lib.binding_parameters[drug]['Kmax']
-Ku = param_lib.binding_parameters[drug]['Ku']
-Hill_n = param_lib.binding_parameters[drug]['N']
-half_effect_conc = param_lib.binding_parameters[drug]['EC50']
-
-all_params = [Vhalf, Kmax, Ku, Hill_n, half_effect_conc]
 
 # Load IKr model
 model = '../../model/ohara-cipa-v1-2017-IKr.mmt'
@@ -58,36 +39,51 @@ save_signal = 2
 repeats = 1000
 APD_points = 20
 
-# Perturbing the parameters
+param_lib = modelling.BindingParameters()
+drug_list = param_lib.drug_compounds
+
 SA_model = modelling.SensitivityAnalysis()
-param_ranges = []
 param_names = SA_model.param_names
+parameter_interest = param_names[:3]
 
-orig_param_values = pd.DataFrame(all_params, index=param_names)
-orig_param_values = orig_param_values.T
-ComparisonController = modelling.ModelComparison(orig_param_values)
 
-for k, parameter_interest in enumerate(param_names):
-    filename = 'SA_' + drug + '_' + parameter_interest + '.csv'
-    if os.path.exists(saved_data_dir + filename):
-        os.remove(saved_data_dir + filename)
+def param_evaluation(param, drug, param_values):
 
-    param_range = SA_model.param_explore_drug(drug, parameter_interest)
-    param_values = orig_param_values
+    Hill_n = param_lib.binding_parameters[drug]['N']
+    half_effect_conc = param_lib.binding_parameters[drug]['EC50']
 
-    for num, param in enumerate(param_range):
+    print('Running for drug: ', drug, ' and parameter value: ', param)
+    orig_half_effect_conc = param_values['EC50'][0]
+    param_values[parameter_interest][0] = param
+    ComparisonController.drug_param_values = param_values
 
-        param_values[parameter_interest][0] = param
-        ComparisonController.drug_param_values = param_values
+    Hill_curve_coefs, drug_conc_Hill, peaks_norm = \
+        ComparisonController.compute_Hill(drug_model, parallel=False)
 
-        Hill_curve_coefs, drug_conc_Hill, peaks_norm = \
-            ComparisonController.compute_Hill(drug_model)
+    # Normalise drug concentration against EC50
+    param_values['EC50'][0] = 1
+    ComparisonController.drug_param_values = param_values
+    drug_conc_AP = 10**np.linspace(np.log10(drug_conc_Hill[1]),
+                                   np.log10(max(drug_conc_Hill)),
+                                   APD_points)
 
-        drug_conc_range = (np.log10(drug_conc_Hill[1]),
-                           np.log10(max(drug_conc_Hill)))
+    Hill_n = param_values['N'][0]
+    drug_conc_AP = [i / (np.power(half_effect_conc, 1 / Hill_n))
+                    for i in drug_conc_AP]
+
+    if isinstance(Hill_curve_coefs, str):
+        Hill_curve_coefs = [float("nan")] * 2
+        APD_trapping = [float("Nan")] * APD_points
+        APD_conductance = [float("Nan")] * APD_points
+        RMSError = float("Nan")
+        MAError = float("Nan")
+    else:
+        # Simulate action potentials
         try:
-            APD_trapping, APD_conductance = ComparisonController.APD_sim(
-                AP_model, Hill_curve_coefs, drug_conc=drug_conc_range)
+            APD_trapping, APD_conductance, drug_conc_AP = \
+                ComparisonController.APD_sim(
+                    AP_model, Hill_curve_coefs, drug_conc=drug_conc_AP,
+                    EAD=True)
 
             RMSError = ComparisonController.compute_RMSE(APD_trapping,
                                                          APD_conductance)
@@ -99,35 +95,75 @@ for k, parameter_interest in enumerate(param_names):
             RMSError = float("Nan")
             MAError = float("Nan")
 
-        # Create dataframe to save results
-        conc_Hill_ind = ['conc_' + str(i) for i, _ in
-                         enumerate(drug_conc_Hill)]
-        drug_conc_AP = 10**np.linspace(*drug_conc_range, APD_points)
-        conc_AP_ind = ['conc_' + str(i) for i, _ in enumerate(drug_conc_AP)]
-        index_dict = {'drug_conc_Hill': conc_Hill_ind,
-                      'peak_current': conc_Hill_ind,
-                      'Hill_curve': ['Hill_coef', 'IC50'],
-                      'param_values': param_names, 'drug_conc_AP': conc_AP_ind,
-                      'APD_trapping': conc_AP_ind,
-                      'APD_conductance': conc_AP_ind, 'RMSE': ['RMSE'],
-                      'MAE': ['MAE']}
-        all_index = [(i, j) for i in index_dict.keys() for j in index_dict[i]]
-        index = pd.MultiIndex.from_tuples(all_index)
+    # Create dataframe to save results
+    conc_Hill_ind = ['conc_' + str(i) for i, _ in
+                     enumerate(drug_conc_Hill)]
+    conc_AP_ind = ['conc_' + str(i) for i, _ in enumerate(drug_conc_AP)]
+    index_dict = {'drug_conc_Hill': conc_Hill_ind,
+                  'peak_current': conc_Hill_ind,
+                  'Hill_curve': ['Hill_coef', 'IC50'],
+                  'param_values': param_names, 'drug_conc_AP': conc_AP_ind,
+                  'APD_trapping': conc_AP_ind,
+                  'APD_conductance': conc_AP_ind, 'RMSE': ['RMSE'],
+                  'MAE': ['MAE']}
+    all_index = [(i, j) for i in index_dict.keys() for j in index_dict[i]]
+    index = pd.MultiIndex.from_tuples(all_index)
 
-        big_df = pd.DataFrame(
-            drug_conc_Hill + list(peaks_norm) + list(Hill_curve_coefs) +
-            list(param_values.values[0]) + list(drug_conc_AP) + APD_trapping +
-            APD_conductance + [RMSError] + [MAError], index=index)
+    param_values['EC50'][0] = orig_half_effect_conc
+    big_df = pd.DataFrame(
+        drug_conc_Hill + list(peaks_norm) + list(Hill_curve_coefs) +
+        list(param_values.values[0]) + list(drug_conc_AP) + APD_trapping +
+        APD_conductance + [RMSError] + [MAError], index=index)
 
-        if num != 0:
-            previous_df = pd.read_csv(saved_data_dir + filename,
+    return big_df
+
+filename = 'SA_all.csv'
+if os.path.exists(saved_data_dir + filename):
+    saved_results_df = pd.read_csv(saved_data_dir + filename,
+                                    header=[0, 1], index_col=[0],
+                                    skipinitialspace=True)
+    ran_values = saved_results_df['param_values'][
+        parameter_interest].values
+else:
+    ran_values = []
+
+
+    Hill_n = param_lib.binding_parameters[drug]['N']
+    half_effect_conc = param_lib.binding_parameters[drug]['EC50']
+
+    all_params = [Vhalf, Kmax, Ku, Hill_n, half_effect_conc]
+
+    orig_param_values = pd.DataFrame(all_params, index=param_names)
+    orig_param_values = orig_param_values.T
+    ComparisonController = modelling.ModelComparison(orig_param_values)
+
+    param_values = orig_param_values
+    param_range = SA_model.param_explore_drug(drug, parameter_interest)
+
+
+    param_range = [i for i in param_range if i not in ran_values]
+    n_workers = 7
+    evaluator = pints.ParallelEvaluator(param_evaluation,
+                                        n_workers=n_workers,
+                                        args=[drug, param_values])
+    for i in range(int(np.ceil(len(param_range) / n_workers))):
+        print('Running samples ', n_workers * i, 'to',
+              n_workers * (i + 1) - 1)
+        big_df = evaluator.evaluate(
+            param_range[i * n_workers: (i + 1) * n_workers])
+
+        if os.path.exists(saved_data_dir + filename):
+            combined_df = pd.read_csv(saved_data_dir + filename,
                                       header=[0, 1], index_col=[0],
                                       skipinitialspace=True)
-            comb_df = pd.concat([previous_df, big_df.T])
+            for i in range(len(big_df)):
+                combined_df = pd.concat([combined_df, big_df[i].T])
         else:
-            comb_df = big_df.T
+            combined_df = big_df[0].T
+            for i in range(1, len(big_df)):
+                combined_df = pd.concat([combined_df, big_df[i].T])
 
-        comb_df.to_csv(saved_data_dir + filename)
+        combined_df.to_csv(saved_data_dir + filename)
 
         os.system('cp ' + saved_data_dir + filename + ' ' +
                   saved_data_dir + filename[:-4] + '_copy.csv')
